KatanaPlugin: Handle USD Lights.
USD lights are translated to Katana lights.

All lights are scanned and put into a lightList at /root/world in Katana
to ensure render find them.

Because UsdLuxLight is not a concrete schema, we made an ArnoldLight
that inherits from UsdLuxLight and allow us to have generic ArnoldLight
into USD scene.